chore(cleaning): Apply review advices
