+++ .vscode/settings.json
@@ -4,5 +4,6 @@
     },
     "eslint.validate": [
         "javascript"
-    ]
+    ],
+    "editor.detectIndentation": false
 }
