+++ AndroidManifest.xml
@@ -1,0 +1,9 @@
+<manifest
+  xmlns:tools="http://schemas.android.com/tools">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+
+    <application
+      android:usesCleartextTraffic="true" tools:targetApi="28">
+    </application>
+</manifest>
