+++ domains/home/MainSreen.jsx
@@ -1,0 +1,13 @@
+import React, { useState } from "react";
+import PlayerInputName from "./PlayerNameInput.jsx";
+import Welcome from "./components/Welcome.jsx";
+
+
+export default function MainScreen({ navigation }) {
+    const [playerName, setPlayerName] = useState("");
+
+    if (!playerName) {
+        return <PlayerInputName setPlayerName={setPlayerName}></PlayerInputName>;
+    }
+    return <Welcome playerName={playerName} navigation={navigation}></Welcome>;
+}
