+++ domains/main.test.jsx
@@ -1,3 +1,3 @@
 import { it } from "@jest/globals";
 
-it("Should make a basic test to init repo", () => {});
+it("Should make a basic test to init repo", () => { });
